reorganization of preprocessor classes
